Moved NotebookService interface into lab directory.
