+++ api/src/main/java/org/pmiops/workbench/notebooks/NotebooksServiceImpl.java
@@ -29,6 +29,7 @@ import org.pmiops.workbench.exceptions.NotImplementedException;
 import org.pmiops.workbench.firecloud.FireCloudService;
 import org.pmiops.workbench.google.CloudStorageClient;
 import org.pmiops.workbench.google.GoogleCloudLocators;
+import org.pmiops.workbench.lab.notebooks.NotebooksService;
 import org.pmiops.workbench.model.AppType;
 import org.pmiops.workbench.model.FileDetail;
 import org.pmiops.workbench.model.KernelTypeEnum;
