+++ api/src/test/java/org/pmiops/workbench/api/WorkspacesControllerTest.java
@@ -173,7 +173,7 @@ import org.pmiops.workbench.model.WorkspaceResource;
 import org.pmiops.workbench.model.WorkspaceResourceResponse;
 import org.pmiops.workbench.model.WorkspaceResponseListResponse;
 import org.pmiops.workbench.model.WorkspaceUserRolesResponse;
-import org.pmiops.workbench.notebooks.NotebooksService;
+import org.pmiops.workbench.lab.notebooks.NotebooksService;
 import org.pmiops.workbench.rawls.model.RawlsWorkspaceACL;
 import org.pmiops.workbench.rawls.model.RawlsWorkspaceACLUpdate;
 import org.pmiops.workbench.rawls.model.RawlsWorkspaceACLUpdateResponseList;
@@ -1770,7 +1770,7 @@ public class WorkspacesControllerTest {
     Map<String, Cohort> cohortsByName = Maps.uniqueIndex(cohorts, c -> c.getName());
     Map<String, CohortReview> cohortReviewsByName =
         Maps.uniqueIndex(cohortReviews, c -> c.getCohortName());
-    assertThat(cohortsByName.keySet().size()).isEqualTo(2);
+    assertThat(cohortsByName.size()).isEqualTo(2);
     assertThat(cohortsByName.keySet()).containsExactly("c1", "c2");
     assertThat(cohorts.stream().map(c -> c.getId()).collect(Collectors.toList()))
         .containsNoneOf(c1.getId(), c2.getId());
