+++ api/src/test/java/org/pmiops/workbench/interactiveanalysis/InteractiveAnalysisServiceTest.java
@@ -23,7 +23,7 @@ import org.pmiops.workbench.exceptions.NotImplementedException;
 import org.pmiops.workbench.firecloud.FireCloudService;
 import org.pmiops.workbench.leonardo.LeonardoApiClient;
 import org.pmiops.workbench.model.AppType;
-import org.pmiops.workbench.notebooks.NotebooksService;
+import org.pmiops.workbench.lab.notebooks.NotebooksService;
 import org.pmiops.workbench.notebooks.model.StorageLink;
 import org.pmiops.workbench.rawls.model.RawlsWorkspaceDetails;
 import org.pmiops.workbench.rawls.model.RawlsWorkspaceResponse;
@@ -61,7 +61,7 @@ public class InteractiveAnalysisServiceTest {
     }
   }
 
-  private static WorkbenchConfig config = WorkbenchConfig.createEmptyConfig();
+  private static final WorkbenchConfig config = WorkbenchConfig.createEmptyConfig();
   private static DbUser user;
   private String aouConfigDataUri;
