+++ api/src/test/java/org/pmiops/workbench/notebooks/NotebooksServiceTest.java
@@ -44,6 +44,7 @@ import org.pmiops.workbench.exceptions.FailedPreconditionException;
 import org.pmiops.workbench.exceptions.NotImplementedException;
 import org.pmiops.workbench.firecloud.FireCloudService;
 import org.pmiops.workbench.google.CloudStorageClient;
+import org.pmiops.workbench.lab.notebooks.NotebooksService;
 import org.pmiops.workbench.model.AppType;
 import org.pmiops.workbench.model.FileDetail;
 import org.pmiops.workbench.model.KernelTypeEnum;
@@ -383,7 +384,7 @@ public class NotebooksServiceTest {
   public void testGetNotebookContents() {
     JSONObject expectedResult = new JSONObject();
     when(mockCloudStorageClient.getBlob(anyString(), anyString())).thenReturn(mockBlob);
-    when(mockBlob.getSize()).thenReturn(1l);
+    when(mockBlob.getSize()).thenReturn(1L);
     when(mockCloudStorageClient.readBlobAsJson(any())).thenReturn(expectedResult);
 
     JSONObject actualResult = notebooksService.getNotebookContents(BUCKET_NAME, "notebookName");
@@ -424,7 +425,7 @@ public class NotebooksServiceTest {
         WorkspaceAccessLevel.OWNER);
     when(mockCloudStorageClient.readBlobAsJson(any())).thenReturn(notebookFile);
     when(mockCloudStorageClient.getBlob(anyString(), anyString())).thenReturn(mockBlob);
-    when(mockBlob.getSize()).thenReturn(1l);
+    when(mockBlob.getSize()).thenReturn(1L);
 
     KernelTypeEnum kernelType =
         notebooksService.getNotebookKernel(
@@ -865,7 +866,7 @@ public class NotebooksServiceTest {
     stubGetWorkspace(NAMESPACE_NAME, WORKSPACE_NAME, BUCKET_NAME, WorkspaceAccessLevel.OWNER);
 
     when(mockCloudStorageClient.getBlob(anyString(), anyString())).thenReturn(mockBlob);
-    when(mockBlob.getSize()).thenReturn(1l);
+    when(mockBlob.getSize()).thenReturn(1L);
     when(mockBlob.getContent()).thenReturn(new byte[10]);
   }
 }
